+++ oalg-base/src/OAlg/Hom/Vectorial.hs
@@ -28,6 +28,8 @@ import OAlg.Prelude
 
 import OAlg.Category.Path
 
+import OAlg.Structure.Fibred.Definition
+import OAlg.Structure.Additive.Definition
 import OAlg.Structure.Vectorial
 
 import OAlg.Hom.Definition
@@ -42,15 +44,20 @@ import OAlg.Hom.Additive
 -- __Property__ Let @__h__@ be a type instance of the class @'HomVectorial' __k__@, then
 -- for all @__a__@, @__b__@, @v@ in @__h__ __a__ __b__@ and @x@ in @__k__@ holds:
 -- @'amap' h (x '!' v) '==' x '!' 'amap' h v@.
-class (EmbeddableMorphism h (Vec k), HomAdditive h) => HomVectorial k h
+class (HomAdditive h, Transformable (ObjectClass h) (Vec k)) => HomVectorial k h
+
 
 instance HomVectorial k h => HomVectorial k (Path h)
 
-instance (ForgetfulVec k s, ForgetfulTyp (s k))
+instance ( TransformableFbr (s k), TransformableTyp (s k)
+         , TransformableAdd (s k)
+         , TransformableVec k s
+         )
   => HomVectorial k (IdHom (s k))
 
 --------------------------------------------------------------------------------
 -- Hom -
 
 type instance Hom (Vec k) h = HomVectorial k h
+